chore(AddAnother): clean up code, move methods into appropriate classes
