+++ src/Helpers/PageHelpers/IPageHelper.cs
@@ -11,6 +11,8 @@ namespace form_builder.Helpers.PageHelpers
 
         (FormSchema dynamicFormSchema, Page dynamicCurrentPage) GetDynamicFormSchema(Page currentPage, string guid);
 
+        void RemoveFieldset(Dictionary<string, dynamic> viewModel, string form, string guid, string path, string removeKey);
+
         void RemoveAnswers(Dictionary<string, dynamic> answersToRemove, string guid, string path);
 
         void SaveAnswers(Dictionary<string, dynamic> viewModel, string guid, string form, IEnumerable<CustomFormFile> files, bool isPageValid, bool appendMultipleFileUploadParts = false);
