+++ src/Models/Properties/ElementProperties/PostbackButtonProperty.cs
@@ -1,14 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace form_builder.Models.Properties.ElementProperties
+﻿namespace form_builder.Models.Properties.ElementProperties
 {
     public partial class BaseProperty
     {
-        public string Action { get; set; }
-        public string Controller { get; set; }
         public string Name { get; set; }
     }
 }
